+++ java/fury-core/src/test/java/io/fury/serializer/ProtocolInteroperabilityTest.java
@@ -27,6 +27,8 @@ import io.fury.FuryTestBase;
 import io.fury.config.CompatibleMode;
 import io.fury.config.Language;
 import io.fury.resolver.MetaContext;
+import io.fury.serializer.collection.CollectionSerializersTest;
+import io.fury.serializer.map.MapSerializersTest;
 import io.fury.test.bean.BeanA;
 import io.fury.test.bean.BeanB;
 import io.fury.test.bean.MapFields;
